requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=1.4.23
pymysql>=1.0.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
alembic>=1.7.1
python-dotenv>=0.19.0
pydantic[email]>=1.8.0, <2.0.0 
